element: recipe: slightly shrink ingredient list
Print view can lead to some weird viewport width etc issues. Currently
the col-8 ingredient section in this component will show lots of
deadspace and sometimes compress the method section. This commit reduces
the sapce of the amount column and subsequently of the whole ingredietn
section to fix this layouting issue.