fix: completely typed tags from list are not accepted
Since our searchbar relies on the html5 `list` attribute of `<input>` tags
and their implementation, we have very limited options to detect when a
user selects an item from the autocomplete list. Currently, we are
relying on capturing `InputEvents` which fire with`inputType=="insertReplacementText"`
when the browser autofills from the input list. This approach is however
limited. Atleast chromium browsers save work by only replacing the text
if it differs from the autocomplete selection. This means that if the
user already fully typed the tag that they wanted, the selection will
cause no event and the tag will never actually be selected. This happens
commonly with very short tags like "gin"

This commit adds a trailing space to every tag, therefore ensuring that
the value selected from the list will almost always differ from what the
user typed. Cocktails get an added cocktail emoji.
This trailing whitespace/emoji is then trimmed when evaluating the entered tag.
Note that his assumes that it is uncommon for the user to type a trailing space
and then select from the list. It is assumed that no user will type an
emoji into the search, making cocktail selection even more robust.